New migration step with consolidated migrations prior to 2.0
