--- lib/Migration/Version0009Date20181125061900.php
@@ -1,366 +1,0 @@
-<?php
-/**
- * @copyright Copyright (c) 2017 René Gieling <github@dartcafe.de>
- *
- * @author René Gieling <github@dartcafe.de>
- *
- * @license GNU AGPL version 3 or any later version
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-namespace OCA\Polls\Migration;
-
-use OCP\DB\ISchemaWrapper;
-use OCP\IConfig;
-use OCP\IDBConnection;
-use OCP\Migration\SimpleMigrationStep;
-use OCP\Migration\IOutput;
-
-/**
- * Installation class for the polls app.
- * Initial db creation
- */
-class Version0009Date20181125061900 extends SimpleMigrationStep {
-
-	/** @var IDBConnection */
-	protected $connection;
-
-	/** @var IConfig */
-	protected $config;
-
-	public function __construct(IDBConnection $connection, IConfig $config) {
-		$this->connection = $connection;
-		$this->config = $config;
-	}
-
-	/**
-	 * $schemaClosure The `\Closure` returns a `ISchemaWrapper`
-	 */
-	public function changeSchema(IOutput $output, \Closure $schemaClosure, array $options) {
-		/** @var ISchemaWrapper $schema */
-		$schema = $schemaClosure();
-
-		if (!$schema->hasTable('polls_events')) {
-			$table = $schema->createTable('polls_events');
-			$table->addColumn('id', 'integer', [
-				'autoincrement' => true,
-				'notnull' => true,
-			]);
-			$table->addColumn('hash', 'string', [
-				'notnull' => false,
-				'length' => 64,
-			]);
-			$table->addColumn('type', 'bigint', [
-				'notnull' => false,
-				'length' => 16,
-			]);
-			$table->addColumn('title', 'string', [
-				'notnull' => true,
-				'length' => 128,
-			]);
-			$table->addColumn('description', 'string', [
-				'notnull' => true,
-				'length' => 1024,
-			]);
-			$table->addColumn('owner', 'string', [
-				'notnull' => true,
-				'length' => 64,
-			]);
-			$table->addColumn('created', 'datetime', [
-				'notnull' => false,
-			]);
-			$table->addColumn('access', 'string', [
-				'notnull' => false,
-				'length' => 1024,
-			]);
-			$table->addColumn('expire', 'datetime', [
-				'notnull' => false,
-			]);
-			$table->addColumn('is_anonymous', 'integer', [
-				'notnull' => false,
-				'default' => 0,
-			]);
-			$table->addColumn('full_anonymous', 'integer', [
-				'notnull' => false,
-				'default' => 0,
-			]);
-			$table->addColumn('allow_maybe', 'integer', [
-				'notnull' => false,
-				'default' => 1,
-			]);
-			$table->setPrimaryKey(['id']);
-		} else {
-			$table = $schema->getTable('polls_events');
-			if (!$table->hasColumn('allow_maybe')) {
-				$table->addColumn('allow_maybe', 'integer', [
-					'notnull' => false,
-					'default' => 1,
-				]);
-			}
-		}
-
-		if (!$schema->hasTable('polls_options')) {
-			$table = $schema->createTable('polls_options');
-			$table->addColumn('id', 'integer', [
-				'autoincrement' => true,
-				'notnull' => true,
-			]);
-			$table->addColumn('poll_id', 'integer', [
-				'notnull' => false,
-			]);
-			$table->addColumn('poll_option_text', 'string', [
-				'notnull' => false,
-				'length' => 256,
-			]);
-			$table->addColumn('timestamp', 'integer', [
-				'notnull' => false,
-				'default' => 0
-			]);
-			$table->setPrimaryKey(['id']);
-		}
-
-		if (!$schema->hasTable('polls_votes')) {
-			$table = $schema->createTable('polls_votes');
-			$table->addColumn('id', 'integer', [
-				'autoincrement' => true,
-				'notnull' => true,
-			]);
-			$table->addColumn('poll_id', 'integer', [
-				'notnull' => false,
-			]);
-			$table->addColumn('user_id', 'string', [
-				'notnull' => true,
-				'length' => 64,
-			]);
-			$table->addColumn('vote_option_id', 'integer', [
-				'notnull' => true,
-				'default' => 0,
-				'length' => 64,
-			]);
-			$table->addColumn('vote_option_text', 'string', [
-				'notnull' => false,
-				'length' => 256,
-			]);
-			$table->addColumn('vote_answer', 'string', [
-				'notnull' => false,
-				'length' => 64,
-			]);
-			$table->setPrimaryKey(['id']);
-		}
-
-		if (!$schema->hasTable('polls_comments')) {
-			$table = $schema->createTable('polls_comments');
-			$table->addColumn('id', 'integer', [
-				'autoincrement' => true,
-				'notnull' => true,
-			]);
-			$table->addColumn('poll_id', 'integer', [
-				'notnull' => false,
-			]);
-			$table->addColumn('user_id', 'string', [
-				'notnull' => true,
-				'length' => 64,
-			]);
-			$table->addColumn('dt', 'string', [
-				'notnull' => true,
-				'length' => 32,
-			]);
-			$table->addColumn('comment', 'string', [
-				'notnull' => false,
-				'length' => 1024,
-			]);
-			$table->setPrimaryKey(['id']);
-		}
-
-		if (!$schema->hasTable('polls_notif')) {
-			$table = $schema->createTable('polls_notif');
-			$table->addColumn('id', 'integer', [
-				'autoincrement' => true,
-				'notnull' => true,
-			]);
-			$table->addColumn('poll_id', 'integer', [
-				'notnull' => false,
-			]);
-			$table->addColumn('user_id', 'string', [
-				'notnull' => true,
-				'length' => 64,
-			]);
-			$table->setPrimaryKey(['id']);
-		}
-
-		return $schema;
-	}
-
-	/**
-	 * $schemaClosure The `\Closure` returns a `ISchemaWrapper`
-	 *
-	 * @return void
-	 */
-	public function postSchemaChange(IOutput $output, \Closure $schemaClosure, array $options) {
-		/** @var ISchemaWrapper $schema */
-		$schema = $schemaClosure();
-
-		if ($schema->hasTable('polls_dts')) {
-			$this->copyDateOptions();
-			$this->copyDateVotes();
-		}
-		if ($schema->hasTable('polls_txts')) {
-			$this->copyTextOptions();
-			$this->copyTextVotes();
-		}
-	}
-
-	/**
-	 * Copy date options
-	 */
-	protected function copyDateOptions(): void {
-		$insert = $this->connection->getQueryBuilder();
-		$insert->insert('polls_options')
-			->values([
-				'poll_id' => $insert->createParameter('poll_id'),
-				'poll_option_text' => $insert->createParameter('poll_option_text'),
-			]);
-		$query = $this->connection->getQueryBuilder();
-		$query->select('*')
-			->from('polls_dts');
-		$result = $query->execute();
-		while ($row = $result->fetch()) {
-			$insert
-				->setParameter('poll_id', $row['poll_id'])
-				->setParameter('poll_option_text', date($row['dt']));
-			$insert->execute();
-		}
-		$result->closeCursor();
-	}
-
-	/**
-	 * Copy text options
-	 */
-	protected function copyTextOptions(): void {
-		$insert = $this->connection->getQueryBuilder();
-		$insert->insert('polls_options')
-			->values([
-				'poll_id' => $insert->createParameter('poll_id'),
-				'poll_option_text' => $insert->createParameter('poll_option_text'),
-			]);
-		$query = $this->connection->getQueryBuilder();
-		$query->select('*')
-			->from('polls_txts');
-		$result = $query->execute();
-		while ($row = $result->fetch()) {
-			$insert
-				->setParameter('poll_id', $row['poll_id'])
-				->setParameter('poll_option_text', preg_replace("/_\d*$/", "$1", $row['text']));
-			$insert->execute();
-		}
-		$result->closeCursor();
-	}
-
-	/**
-	 * Copy date votes
-	 */
-	protected function copyDateVotes(): void {
-		$insert = $this->connection->getQueryBuilder();
-		$insert->insert('polls_votes')
-			->values([
-				'poll_id' => $insert->createParameter('poll_id'),
-				'user_id' => $insert->createParameter('user_id'),
-				'vote_option_id' => $insert->createParameter('vote_option_id'),
-				'vote_option_text' => $insert->createParameter('vote_option_text'),
-				'vote_answer' => $insert->createParameter('vote_answer'),
-			]);
-		$query = $this->connection->getQueryBuilder();
-		$query->select('*')
-			->from('polls_particip');
-		$result = $query->execute();
-		while ($row = $result->fetch()) {
-			$insert
-				->setParameter('poll_id', $row['poll_id'])
-				->setParameter('user_id', $row['user_id'])
-				->setParameter('vote_option_id', $this->findOptionId($row['poll_id'], $row['dt']))
-				->setParameter('vote_option_text', date($row['dt']))
-				->setParameter('vote_answer', $this->translateVoteTypeToAnswer($row['type']));
-			$insert->execute();
-		}
-		$result->closeCursor();
-	}
-
-	/**
-	 * Copy text votes
-	 */
-	protected function copyTextVotes(): void {
-		$insert = $this->connection->getQueryBuilder();
-		$insert->insert('polls_votes')
-			->values([
-				'poll_id' => $insert->createParameter('poll_id'),
-				'user_id' => $insert->createParameter('user_id'),
-				'vote_option_id' => $insert->createParameter('vote_option_id'),
-				'vote_option_text' => $insert->createParameter('vote_option_text'),
-				'vote_answer' => $insert->createParameter('vote_answer'),
-			]);
-		$query = $this->connection->getQueryBuilder();
-		$query->select('*')
-			->from('polls_particip_text');
-		$result = $query->execute();
-		while ($row = $result->fetch()) {
-			$insert
-				->setParameter('poll_id', $row['poll_id'])
-				->setParameter('user_id', $row['user_id'])
-				->setParameter('vote_option_id', $this->findOptionId($row['poll_id'], preg_replace("/_\d*$/", "$1", $row['text'])))
-				->setParameter('vote_option_text', preg_replace("/_\d*$/", "$1", $row['text']))
-				->setParameter('vote_answer', $this->translateVoteTypeToAnswer($row['type']));
-			$insert->execute();
-		}
-		$result->closeCursor();
-	}
-
-	protected function findOptionId(int $pollId, ?string $text): int {
-		$queryFind = $this->connection->getQueryBuilder();
-		$queryFind->select(['id'])
-			->from('polls_options')
-			->where('poll_id = :pollId')
-			->andWhere('poll_option_text = :text')
-			->setParameter('pollId', $pollId)
-			->setParameter('text', $text);
-
-		$resultFind = $queryFind->execute();
-		$row = $resultFind->fetch();
-
-		if ($row['id'] === null) {
-			// seems, we have an orphaned vote, return 0 as vote_option_id
-			return 0;
-		} else {
-			return $row['id'];
-		}
-	}
-
-	protected function translateVoteTypeToAnswer(string $voteType): string {
-		switch ($voteType) {
-			case 0:
-				$answer = "no";
-				break;
-			case 1:
-				$answer = "yes";
-				break;
-			case 2:
-				$answer = "maybe";
-				break;
-			default:
-				$answer = "no";
-		}
-		return $answer;
-	}
-}
