--- lib/Migration/Version0009Date20181125062101.php
@@ -1,55 +1,0 @@
-<?php
-/**
- * @copyright Copyright (c) 2017 Joas Schilling <coding@schilljs.com>
- *
- * @author Joas Schilling <coding@schilljs.com>
- *
- * @license GNU AGPL version 3 or any later version
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-namespace OCA\Polls\Migration;
-
-use OCP\DB\ISchemaWrapper;
-use OCP\Migration\SimpleMigrationStep;
-use OCP\Migration\IOutput;
-
-/**
- * Deleting unused tables 'polls_dts', 'polls_txts', 'polls_particip' and 'polls_particip_text'
- * after migration in 'Version000009Date20171202105141.php'
- */
-
-class Version0009Date20181125062101 extends SimpleMigrationStep {
-	/**
-	 * $schemaClosure The `\Closure` returns a `ISchemaWrapper`
-	 */
-	public function changeSchema(IOutput $output, \Closure $schemaClosure, array $options) {
-		/** @var ISchemaWrapper $schema */
-		$schema = $schemaClosure();
-		if ($schema->hasTable('polls_dts')) {
-			$schema->dropTable('polls_dts');
-		}
-		if ($schema->hasTable('polls_txts')) {
-			$schema->dropTable('polls_txts');
-		}
-		if ($schema->hasTable('polls_particip')) {
-			$schema->dropTable('polls_particip');
-		}
-		if ($schema->hasTable('polls_particip_text')) {
-			$schema->dropTable('polls_particip_text');
-		}
-		return $schema;
-	}
-}
