--- lib/Migration/Version0108Date20210307130001.php
@@ -1,114 +1,0 @@
-<?php
-/**
- * @copyright Copyright (c) 2017 René Gieling <github@dartcafe.de>
- *
- * @author René Gieling <github@dartcafe.de>
- *
- * @license GNU AGPL version 3 or any later version
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-namespace OCA\Polls\Migration;
-
-use OCP\DB\ISchemaWrapper;
-use OCP\IConfig;
-use OCP\IDBConnection;
-use OCP\Migration\SimpleMigrationStep;
-use OCP\Migration\IOutput;
-
-class Version0108Date20210307130001 extends SimpleMigrationStep {
-
-	/** @var IDBConnection */
-	protected $connection;
-
-	/** @var IConfig */
-	protected $config;
-
-	public function __construct(IDBConnection $connection, IConfig $config) {
-		$this->connection = $connection;
-		$this->config = $config;
-	}
-
-	public function changeSchema(IOutput $output, \Closure $schemaClosure, array $options) {
-		/** @var ISchemaWrapper $schema */
-		$schema = $schemaClosure();
-
-		if ($schema->hasTable('polls_options')) {
-			$table = $schema->getTable('polls_options');
-			foreach ($table->getIndexes() as $index) {
-				if (strpos($index->getName(), 'UNIQ_') === 0) {
-					$table->dropIndex($index->getName());
-				}
-			}
-		}
-
-		if ($schema->hasTable('polls_log')) {
-			$table = $schema->getTable('polls_log');
-			foreach ($table->getIndexes() as $index) {
-				if (strpos($index->getName(), 'UNIQ_') === 0) {
-					$table->dropIndex($index->getName());
-				}
-			}
-		}
-
-		if ($schema->hasTable('polls_notif')) {
-			$table = $schema->getTable('polls_notif');
-			foreach ($table->getIndexes() as $index) {
-				if (strpos($index->getName(), 'UNIQ_') === 0) {
-					$table->dropIndex($index->getName());
-				}
-			}
-		}
-
-		if ($schema->hasTable('polls_share')) {
-			$table = $schema->getTable('polls_share');
-			foreach ($table->getIndexes() as $index) {
-				if (strpos($index->getName(), 'UNIQ_') === 0) {
-					$table->dropIndex($index->getName());
-				}
-			}
-		}
-
-		if ($schema->hasTable('polls_votes')) {
-			$table = $schema->getTable('polls_votes');
-			foreach ($table->getIndexes() as $index) {
-				if (strpos($index->getName(), 'UNIQ_') === 0) {
-					$table->dropIndex($index->getName());
-				}
-			}
-		}
-
-		if ($schema->hasTable('polls_preferences')) {
-			$table = $schema->getTable('polls_preferences');
-			foreach ($table->getIndexes() as $index) {
-				if (strpos($index->getName(), 'UNIQ_') === 0) {
-					$table->dropIndex($index->getName());
-				}
-			}
-		}
-
-		if ($schema->hasTable('polls_watch')) {
-			$table = $schema->getTable('polls_watch');
-			foreach ($table->getIndexes() as $index) {
-				if (strpos($index->getName(), 'UNIQ_') === 0) {
-					$table->dropIndex($index->getName());
-				}
-			}
-		}
-
-		return $schema;
-	}
-}
